Reimplement the KisSliderSpinBox and KisDoubleSliderSpinBox widgets

+++ libs/ui/widgets/KisSelectionPropertySlider.cpp
@@ -36,7 +36,7 @@ void KisSelectionPropertySliderBase::setPrefixes(const QString &normalPrefix, co
     setPrefix(normalPrefix);
 }
 
-void KisSelectionPropertySliderBase::setInternalValue(int _value, bool blockUpdateSignal)
+void KisSelectionPropertySliderBase::setInternalValue(qreal _value, bool blockUpdateSignal)
 {
     static const qreal eps = 1e-3;
